@@ -10,8 +10,8 @@ let open = require("amqplib").connect("amqp://192.168.0.207");
 open.then(function (conn) {
 	return conn.createChannel();
 }).then(function (ch) {
-	return ch.assertQueue(q).then(function (ok) {
-		return ch.sendToQueue(q, new Buffer("something to do"));
+	return ch.assertQueue(q).then(function () {
+		return ch.sendToQueue(q, Buffer.from("something to do"));
 	});
 }).catch(console.warn);
 
