@@ -150,7 +150,7 @@ describe("Test AmqpTransporter subscribe", () => {
 					.toHaveBeenCalledWith("MOL-TEST.RES.node", jasmine.any(Function), { noAck: true });
 
 				const consumeCb = transporter.channel.consume.mock.calls[0][1];
-				consumeCb({ content: new Buffer("data") });
+				consumeCb({ content: Buffer.from("data") });
 
 				expect(msgHandler).toHaveBeenCalledTimes(1);
 				expect(transporter.channel.ack).toHaveBeenCalledTimes(0);
@@ -169,7 +169,7 @@ describe("Test AmqpTransporter subscribe", () => {
 					.toHaveBeenCalledWith("MOL-TEST.INFO.node", jasmine.any(Function), { noAck: true });
 
 				const consumeCb = transporter.channel.consume.mock.calls[0][1];
-				consumeCb({ content: new Buffer("data") });
+				consumeCb({ content: Buffer.from("data") });
 
 				expect(msgHandler).toHaveBeenCalledTimes(1);
 				expect(transporter.channel.ack).toHaveBeenCalledTimes(0);
@@ -215,7 +215,7 @@ describe("Test AmqpTransporter subscribe", () => {
 					.toHaveBeenCalledWith("MOL-TEST.REQ.example1.testing", jasmine.any(Function));
 
 				const consumeCb = transporter.channel.consume.mock.calls[0][1];
-				consumeCb({ content: new Buffer("data") });
+				consumeCb({ content: Buffer.from("data") });
 
 				expect(msgHandler).toHaveBeenCalledTimes(1);
 				expect(transporter.channel.ack).toHaveBeenCalledTimes(1);
@@ -244,7 +244,7 @@ describe("Test AmqpTransporter subscribe", () => {
 					);
 
 				const consumeCb = transporter.channel.consume.mock.calls[0][1];
-				consumeCb({ content: new Buffer("data") });
+				consumeCb({ content: Buffer.from("data") });
 
 				expect(msgHandler).toHaveBeenCalledTimes(1);
 				expect(transporter.channel.ack).toHaveBeenCalledTimes(0);
@@ -274,7 +274,7 @@ describe("Test AmqpTransporter subscribe", () => {
 						);
 
 					const consumeCb = transporter.channel.consume.mock.calls[0][1];
-					consumeCb({ content: new Buffer("data") });
+					consumeCb({ content: Buffer.from("data") });
 
 					expect(msgHandler).toHaveBeenCalledTimes(1);
 					expect(transporter.channel.ack).toHaveBeenCalledTimes(0);
